add option for simple and complex entry
